+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## Changelog
 
+### 1.3.0
+
+- **New:** Rewrote code to ES6. ([#412](https://github.com/wenzhixin/multiple-select/pull/412)).
+- **New:** Added `filterGroup` option ([#429](https://github.com/wenzhixin/multiple-select/pull/429)).
+- **New:** Added `babel` and refined the code ([#413](https://github.com/wenzhixin/multiple-select/pull/413)).
+- **New:** Used `dist` to output the build files ([#414](https://github.com/wenzhixin/multiple-select/pull/414)).
+- **New:** Added full examples ([#429](https://github.com/wenzhixin/multiple-select/pull/429)).
+- **Update:** Fixed css escape bug ([#423](https://github.com/wenzhixin/multiple-select/pull/423)).
+
 ### 1.2.3
 
 - **New:** Added `onAfterCreate` event ([#196](https://github.com/wenzhixin/multiple-select/pull/196)).